apps/api/pyproject.toml
[project]
name = "malaflow-api"
version = "0.1.0"
description = "FastAPI and MCP backend for MalaFlow MVP"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "bcrypt>=4.1.3",
  "fastapi>=0.115.0",
  "itsdangerous>=2.2.0",
  "mcp[cli]>=1.12.0",
  "openai>=1.99.0",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.4.0",
  "pyjwt>=2.9.0",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.32",
  "twilio>=9.3.7",
  "uvicorn[standard]>=0.30.6"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.2"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]
